+++ src/app/data-manager/page.tsx
@@ -2,6 +2,7 @@
 
 import { DataManagerToolbar } from '@/components/layout';
 import { CPTToast } from '@cpt-group/cpt-prime-react';
+import { ClassMemberDataTable } from '@/components/pages/DataManager';
 import { useRef } from 'react';
 import type { Toast } from 'primereact/toast';
 
@@ -13,7 +14,7 @@ export default function DataManagerPage() {
       <CPTToast ref={toast} />
       <DataManagerToolbar toastRef={toast} />
       <div className="flex-1 p-4">
-        {/* Main content area - to be developed */}
+        <ClassMemberDataTable />
       </div>
     </div>
   );