+++ src/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './caseList';
+export * from './TEST_DATA';